fix(core): Box::transform の corner 計算で i32 オーバーフローを避ける
`self.x + shiftx` を i32 で加算してから f64 にキャストしていたため、
極端な shift で debug ビルドは panic、release は wrap する。C は
l_int32 で加算しており同条件では未定義動作になる。

f64 に広げてから加算する。正常な範囲では結果は変わらない。

Refs: https://github.com/tagawa0525/leptonica-rs/pull/446#discussion_r3814445894

+++ src/core/box_/transform.rs
@@ -42,9 +42,11 @@ impl Box {
             return Box::new_unchecked(0, 0, 0, 0);
         }
         // C evaluates these with `0.5` as a double literal, so the products
-        // are promoted to double before truncation.
-        let x = (scalex as f64 * (self.x + shiftx) as f64 + 0.5).max(0.0) as i32;
-        let y = (scaley as f64 * (self.y + shifty) as f64 + 0.5).max(0.0) as i32;
+        // are promoted to double before truncation. The shift is added in f64
+        // so an extreme offset cannot overflow the i32 step (C adds as
+        // l_int32, which would be undefined behaviour there).
+        let x = (scalex as f64 * (self.x as f64 + shiftx as f64) + 0.5).max(0.0) as i32;
+        let y = (scaley as f64 * (self.y as f64 + shifty as f64) + 0.5).max(0.0) as i32;
         let w = (scalex as f64 * self.w as f64 + 0.5).max(1.0) as i32;
         let h = (scaley as f64 * self.h as f64 + 0.5).max(1.0) as i32;
         Box::new_unchecked(x, y, w, h)
